Get rid of a now broken optimization
With changes elsewhere in the code revolving around the adding and
removing of nodes to the document to bring it up to spec, this
optimization has now become risky and needs to be implemented on Node
for it to work effectively.  Remove the optimization for now as it was
never implemented on the Document and DocumentFragment interfaces.