Implement an autoloader to handle dependencies
This restructures the folders for html elements and future elements of
other types.  It also moves each exception in to its own file to work
with the new autoloader.  It also removes all manual management of
dependencies from all files and makes adjustments to account for
namespace changes due to the restructuring.