Rename mutationMethodMacro to coincide with spec changes
The mutation method macro steps were renamed to converting nodes to node
in the spec.  This renames the method to more closely match the latest
spec.  It also makes the method static.  It also matches latest spec
changes in that it should rethrow any errors while inserting nodes into
a DocumentFragment.