Update Utils's toString method
This method attempts to convert a variable to a string, but will return
the original value passed if it is not a string.  This is needed since
in PHP booleans are represented as 1 and 0 in string form and the
is_string method does not implicitly call an objects __toString method,
if it has one.  Additionally, PHP represents null as the empty string,
however, most times JavaScript will serialize null to the string 'null'.