Fix Range::insertNode() errors
The problem is that we were using Range's public API to update the
range's start and end node/offsets. The problem with this is that the
node tree mutation algorithims update the range before any nodes
are inserted/removed from the tree. This caused issues with the public
API trying to validate the updated information on a tree that had not
yet been modified causing failures. So, add 2 new internal methods to
Range to update these properties without the validation.