Improve performance with many Range and NodeIterator objects
On PHP 7.4+ we use WeakReference to track all these objects allowing
them to be garbage collected. Also, when a test is finished with it's
Window object, we prune all the ranges and
and node iterators that are associated with the Window's document, which
significantly speeds up the tests on PHP versions less than PHP 7.4.