Compare DOM positions without using a TreeWalker
Walking the entire tree looking for another node is extremely
slow on large, deep documents.

Instead traverse up the tree to a common ancestor and compare
which node is a sibling of the other.

Fix setStartOrEnd to compare root nodes before calling computePosition
so that we never try to compare nodes in different documents.