Work around spec bug in xml fragment serializer
As-is, the results of the spec produce the outerHTML of the node
instead of the innerHTML, when using the innerHTML getter, and
the outerHTML of the faux wrapping node when using the outerHTML getter.
This makes it so that we serialize the children of the context node, which
gives the desired result.