Fix parsing named character references
The spec text is a bit ambiguous, but what its asking for is to
consume tokens while the buffer begins with one of the named
character references. The previous implementation had difficulty doing
that, as it was trying to match against complete names.

Instead, we turn the entities.json file into a trie, to make partial
matches easier.